:package:문혜림: 테스트를 위한 더미코드에서 서버와 통신할 수 있는 코드로 수정
